+++ plugins/flatlaf-intellij-themes/src/main/java/is/codion/plugin/flatlaf/intellij/IntelliJThemeProvider.java
@@ -286,8 +286,8 @@ public final class IntelliJThemeProvider implements LookAndFeelProvider {
 		@Override
 		public void accept(LookAndFeelInfo lookAndFeelInfo) {
 			try {
-				Scaler.instance(lookAndFeelInfo.getClassName()).ifPresent(Scaler::apply);
-				UIManager.setLookAndFeel(requireNonNull(lookAndFeelInfo).getClassName());
+				Scaler.instance(requireNonNull(lookAndFeelInfo).getClassName()).ifPresent(Scaler::apply);
+				UIManager.setLookAndFeel(lookAndFeelInfo.getClassName());
 				Utilities.updateComponentTreeForAllWindows();
 			}
 			catch (Exception e) {